requirements.txt
numpy==1.15.2
scipy==1.3.3
torch==1.2.0
torchvision==0.4.0
matplotlib==3.0.0
opencv_python==4.1.2.30
Cython==0.29.16
easydict==1.9
ipython==7.13.0
Pillow==7.1.1

Package           Version
----------------- ---------
backcall          0.2.0
certifi           2021.10.8
cycler            0.11.0
Cython            0.29.16
decorator         5.1.1
easydict          1.9
flatbuffers       2.0
ipython           7.13.0
jedi              0.18.1
kiwisolver        1.3.2
libs              0.0.10
matplotlib        3.0.0
numpy             1.16.6
onnx              1.8.1
onnxruntime       1.10.0
opencv-python     4.1.2.30
parso             0.8.3
pexpect           4.8.0
pickleshare       0.7.5
Pillow            7.1.1
pip               21.2.2
prompt-toolkit    3.0.28
protobuf          3.19.4
ptyprocess        0.7.0
pycocotools       2.0.4
Pygments          2.11.2
pyparsing         3.0.7
python-dateutil   2.8.2
scipy             1.3.3
setuptools        58.0.4
six               1.16.0
torch             1.7.1
torchvision       0.8.2
traitlets         5.1.1
typing_extensions 4.1.1
wcwidth           0.2.5
wheel             0.37.1

